5.6.32 uiSlider support for decimals, VST3 export for macOS
* NEW From now on, uisliders will return double float type values to manage decimals. If required values are integers,  new LUA function has been added : getValueInt()
getModulatorInt(), getModulatorValueInt(), getMinModulatorValueInt(), getMaxModulatorValueInt(), getValueMappedInt(), getValueNonMappedInt(), getMinMappedInt(), getMaxMappedInt()

* NEW VST3 can now export instances without the need to compile intermediate plugins from the Projucer. VST3 identifiers will be taken from the panel plugin name, plugin ID, Panel Author and manufacturer ID as well as the VST3 plugin type. Exported VST3 plugins are codesigned on export directly with JUCE childProcess() automatically, either with local ad-hoc signature or, if selected, with a developer certificate.

* UPDATED Decimal value are now supported for Sliders (ie. 3.1416). CtrlrModulator.cpp & h, CtrlrModulatorProcessor.cpp & h, CtrlrIDs.xml, CtrlrLuaManager.cpp, CtrlrLuaMethodManager.h, CtrlrLuaMethodManagerCalls.cpp, CtrlrFixedImageSlider.cpp, CtrlrFixedImageSlider.cpp, CtrlrImageSlider.cpp, CtrlrSliderInternal.cpp & h
* UPDATED Decimal interval steps are now supported for Sliders (ie. 0.1)
* FIXED Useless shortcuts such as "New Panel", "Export" enabled on restricted instances. CtrlrEditorApplicationCommands.cpp
* FIXED FileChooser still hanging when exporting instance on OSX Catalina & macOS BigSur. CtrlrManager.cpp
* ADDED Slider value Suffix (ie. Hz, ms, dB, etc)
* ADDED JUCE systemStats support for WIN11, macOS 11, macOS 12, macOS 13, macOS 14, macOS 15 and macOS 16. juce_mac_systemStats.cpp, juce_win32_systemStats.cpp, juce_systemStats.h

+++ Source/Lua/Methods/CtrlrLuaMethodManagerCalls.cpp
@@ -107,7 +107,7 @@ const bool CtrlrLuaMethodManager::call(CtrlrLuaMethod *o, CtrlrModulator *param1
 	return (true);
 }
 
-const bool CtrlrLuaMethodManager::call(CtrlrLuaMethod *o, CtrlrModulator *param1, const int param2, const uint8 param3)
+const bool CtrlrLuaMethodManager::call(CtrlrLuaMethod *o, CtrlrModulator *param1, const double param2, const uint8 param3) // Updated v5.6.32. int to double
 {
 	const ScopedLock sl(methodManagerCriticalSection);
